infra/lambda_auth/requirements.txt
google-auth
boto3
requests